### 2025-05-12 #8194 @editorialbot
[PRE REVIEW]: Purify: An R package for resampling and stratified data
**Submitting author:** @jrvanderdoes (<a href="http://orcid.org/0009-0001-9885-3073">Jeremy VanderDoes</a>)
**Repository:** https://github.com/jrvanderdoes/purify
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
