### 2025-06-25 #8194 @spholmes
Hi @simonpcouch: could you help us out by doing a review of this Submission to JOSS?
Thanks
Susan
