### 2025-06-25 #8194 @spholmes
Hi @prdm0, could you help us review this paper for JOSS please.
Many thanks
Susan
