### 2025-07-18 #8194 @spholmes
Hi @T0ngChen,
Are you interested in reviewing this sampling package (Purify) that was submitted to JOSS,
it's a simple checklist process.
Many thanks
Susan
