### 2025-08-03 #8194 @Pentaonia
Hey @spholmes I'm sorry, but I can't make it at the moment. But good luck with the exciting package and the publication!
