### 2025-11-04 #9349 @spholmes
Hi @pachadotdev:
Please start your checklist by typing: `@editorialbot generate my checklist`, especially so we can check the conflict of interest.
Many thanks for your help,
Susan
