### 2025-11-07 #9349 @pachadotdev
> Hi [@pachadotdev](https://github.com/pachadotdev): Please start your checklist by typing: `@editorialbot generate my checklist`, especially so we can check the conflict of interest. Many thanks for your help, Susan

I ran `devtools::check()` and I see these warnings/notes

```
── R CMD check ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
─  using log directory ‘/tmp/RtmpsE66aT/file54e8739e6d6c4/purify.Rcheck’
─  using R version 4.5.1 (2025-06-13)
─  using platform: x86_64-pc-linux-gnu
─  R was compiled by
       gcc (GCC) 15.1.1 20250425
       GNU Fortran (GCC) 15.1.1 20250425
─  running under: Manjaro Linux
─  using session charset: UTF-8
─  using options ‘--no-manual --as-cran’
✔  checking for file ‘purify/DESCRIPTION’ ...
─  this is package ‘purify’ version ‘1.1.0’
─  package encoding: UTF-8
✔  checking package namespace information ...
─  checking package dependencies ... INFO (2s)
   Vignette dependency required without any vignettes: ‘knitr’
✔  checking if this is a source package
✔  checking if there is a namespace
✔  checking for executable files ...
✔  checking for hidden files and directories ...
✔  checking for portable file names
✔  checking for sufficient/correct file permissions ...
✔  checking serialization versions ...
✔  checking whether package ‘purify’ can be installed (3.3s)
✔  checking installed package size ...
✔  checking package directory
✔  checking for future file timestamps (444ms)
✔  checking DESCRIPTION meta-information (606ms)
N  checking top-level files ...
   Non-standard files/directories found at top level:
     ‘paper.bib’ ‘paper.md’
✔  checking for left-over files
✔  checking index information
✔  checking package subdirectories ...
✔  checking code files for non-ASCII characters ...
✔  checking R files for syntax errors ...
✔  checking whether the package can be loaded ...
✔  checking whether the package can be loaded with stated dependencies ...
✔  checking whether the package can be unloaded cleanly ...
✔  checking whether the namespace can be loaded with stated dependencies ...
✔  checking whether the namespace can be unloaded cleanly ...
✔  checking loading without being on the library search path ...
✔  checking dependencies in R code (2.4s)
✔  checking S3 generic/method consistency ...
✔  checking replacement functions ...
✔  checking foreign function calls ...
✔  checking R code for possible problems (3s)
✔  checking Rd files ...
✔  checking Rd metadata ...
✔  checking Rd line widths ...
✔  checking Rd cross-references ...
✔  checking for missing documentation entries ...
✔  checking for code/documentation mismatches ...
✔  checking Rd \usage sections (341ms)
✔  checking Rd contents ...
✔  checking for unstated dependencies in examples ...
✔  checking contents of ‘data’ directory ...
✔  checking data for non-ASCII characters ...
✔  checking LazyData
✔  checking data for ASCII and uncompressed saves ...
─  checking examples ... [15s/15s] OK (15.2s)
✔  checking for unstated dependencies in ‘tests’ ...
─  checking tests ...
   Last 13 lines of output:
     ── Failure ('test-plots.R:4:3'): Plot Strata ───────────────────────────────────
     class(tmp)[1] (`actual`) not equal to "gg" (`expected`).

     `actual`:   "ggplot2::ggplot"
     `expected`: "gg"
     ── Failure ('test-plots.R:7:3'): Plot Strata ───────────────────────────────────
     class(tmp)[1] (`actual`) not equal to "gg" (`expected`).

     `actual`:   "ggplot2::ggplot"
     `expected`: "gg"

     [ FAIL 2 | WARN 0 | SKIP 0 | PASS 54 ]
     Error: Test failures
     Execution halted
✔  checking for non-standard things in the check directory
✔  checking for detritus in the temp directory

   See
     ‘/tmp/RtmpsE66aT/file54e8739e6d6c4/purify.Rcheck/00check.log’
   for details.

── R CMD check results ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────── purify 1.1.0 ────
Duration: 48.8s

❯ checking tests ...
  See below...

❯ checking top-level files ... NOTE
  Non-standard files/directories found at top level:
    ‘paper.bib’ ‘paper.md’

── Test failures ────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────── testthat ────

> # This file is part of the standard setup for testthat.
> # It is recommended that you do not modify it.
> #
> # Where should you do additional test configuration?
> # Learn more about the roles of various files in:
> # * https://r-pkgs.org/tests.html
> # * https://testthat.r-lib.org/reference/test_package.html#special-files
>
> library(testthat)
> library(purify)
>
> test_check("purify")
[ FAIL 2 | WARN 0 | SKIP 0 | PASS 54 ]

── Failure ('test-plots.R:4:3'): Plot Strata ───────────────────────────────────
class(tmp)[1] (`actual`) not equal to "gg" (`expected`).

`actual`:   "ggplot2::ggplot"
`expected`: "gg"
── Failure ('test-plots.R:7:3'): Plot Strata ───────────────────────────────────
class(tmp)[1] (`actual`) not equal to "gg" (`expected`).

`actual`:   "ggplot2::ggplot"
`expected`: "gg"

[ FAIL 2 | WARN 0 | SKIP 0 | PASS 54 ]
Error: Test failures
Execution halted

1 error ✖ | 0 warnings ✔ | 1 note ✖
```

I also had to install forecast as a separate step or it fails
