### 2025-11-10 #9349 @spholmes
Thanks @pachadotdev for starting your review, let's ask the author,  @jrvanderdoes to weigh in on these test failures.
If you could also try other examples  and documentation that are not part of the tests and report on the other parts, that would
keep things moving,
best wishes
Susan
