### 2026-01-01 jrvanderdoes/purify#16 @T0ngChen
[joss] documentation
Per the criteria outlined in the ongoing JOSS review https://github.com/openjournals/joss-reviews/issues/9349#issuecomment-3515260136, there are a few changes that would help address the documentation-related checklist items:

1. **Statement of need**

The software documentation should contain a statement of need, specifically

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

The software documentation does not currently contain a clear and explicit statement of need. It seems that the package's `README` would be a good place to do this. This would also be an appropriate place to explain when and why this package should be used, relative to existing alternatives. It would also be helpful to provide a link in the README to the examples on the package website.

2. **Automated tests**

When running `devtools::check()`, I observed some failing tests. Please fix them.

3. **Community guidelines**
per the requirement : the software documentation should contain a clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

There are currently no explicit guidelines describing how third parties can contribute to the project, report issues, or seek support (e.g., via a CONTRIBUTING file or a brief section in the documentation). Adding such guidance would help clarify these processes for new users and potential contributors.
