### 2026-01-23 #9349 @pachadotdev
> Hi [@jrvanderdoes](https://github.com/jrvanderdoes), do you have some responses about the comments from [@pachadotdev](https://github.com/pachadotdev) and [@T0ngChen](https://github.com/T0ngChen) , there seems to be some elements that are still missing, could you work through the checklists and make the few changes that seem needed. Many Thanks Susan

thanks a lot!

0 rush on my side, I prefer a proper work than a rushed work
